feat: consolidate topbar into menu components, simplify structure, code cleanup, remove store instances
